+++ aspnet-core/Todo.DemoPlugin/Todo.DemoPlugin.Application/Calculation/ICalculationService.cs
@@ -4,6 +4,6 @@ namespace Todo.DemoPlugin.Calculation
 {
     public interface ICalculationService : IApplicationService
     {
-        int Calculation_Add(int x, int y);
+        double Calculation_Add(double x, double y);
     }
 }
